+++ src/platform.ts
@@ -3,6 +3,7 @@ import axios, { AxiosInstance, AxiosRequestConfig } from 'axios';
 import { PLATFORM_NAME, PLUGIN_NAME, DeviceURL } from './settings';
 import { Humidifier } from './Devices/Humidifier';
 import { Bot } from './Devices/Bot';
+import { Meter } from './Devices/Meter';
 import { Curtain } from './Devices/Curtain';
 import { irdevices, device, SwitchBotPlatformConfig, deviceResponses, deviceStatusResponse } from './configTypes';
 
@@ -95,10 +96,16 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         this.config.options.bot.device_press;
         this.config.options.bot.device_switch;
       }
+      
+      // Meter Config Options
+      if (this.config.options?.meter) {
+        this.config.options.meter.hide_temperature;
+        this.config.options.meter.hide_humidity;
+      }
 
       // Humidifier Config Options
       if (this.config.options?.humidifier) {
-        this.config.options.humidifier.hide_tempeture;
+        this.config.options.humidifier.hide_temperature;
       }
 
       // Curtain Config Options
@@ -137,7 +144,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       } else {
         this.log.debug(JSON.stringify(devices));
       }
-      this.log.info(`Total Devices Found: ${devices.body.deviceList.length}`);
+      this.log.info('Total Devices Found:', devices.body.deviceList.length);
       for (const device of devices.body.deviceList) {
         if (this.config.devicediscovery) {
           this.deviceInfo(device);
@@ -154,13 +161,18 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
             this.log.info('Discovered %s %s', device.deviceName, device.deviceType);
             this.createBot(device, devices);
             break;
+          case 'Meter':
+            this.log.info('Discovered %s %s', device.deviceName, device.deviceType);
+            this.createMeter(device, devices);
+            break;
           case 'Curtain':
             this.log.info('Discovered %s %s', device.deviceName, device.deviceType);
             this.createCurtain(device, devices);
             break;
           default:
             this.log.info(
-              `A SwitchBot Device has been discovered with Device Type: ${device.deviceType}, which is currently not supported.`,
+              'A SwitchBot Device has been discovered with Device Type: %s, which is currently not supported.',
+              device.deviceType,
               'Submit Feature Requests Here: https://git.io/JL14Z',
             );
         }
@@ -183,7 +195,11 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
     if (existingAccessory) {
       // the accessory already exists
       if (!this.config.options?.hide_device.includes(device.deviceId) && devices.statusCode === 100) {
-        this.log.info('Restoring existing accessory from cache:', existingAccessory.displayName, 'DeviceID:', device.deviceId);
+        this.log.info(
+          'Restoring existing accessory from cache: %s DeviceID: %s',
+          existingAccessory.displayName,
+          device.deviceId,
+        );
 
         // if you need to update the accessory.context then you should run `api.updatePlatformAccessories`. eg.:
         //existingAccessory.context.firmwareRevision = firmware;
@@ -199,8 +215,12 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       }
     } else if (!this.config.options?.hide_device.includes(device.deviceId)) {
       // the accessory does not yet exist, so we need to create it
-      this.log.info('Adding new accessory:', device.deviceName, device.deviceType, 'DeviceID:', device.deviceId);
-      this.log.debug('Registering new device:', device.deviceName, device.deviceType, '-', device.deviceId);
+      this.log.info(
+        'Adding new accessory: %s %s DeviceID: %s',
+        device.deviceName,
+        device.deviceType,
+        device.deviceId,
+      );
 
       // create a new accessory
       const accessory = new this.api.platformAccessory(`${device.deviceName} ${device.deviceType}`, uuid);
@@ -221,7 +241,12 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
       this.accessories.push(accessory);
     } else {
-      this.log.error(`Unable to Register new device: ${device.deviceName} ${device.deviceType} - ${device.deviceId}`);
+      this.log.error(
+        'Unable to Register new device: %s %s - %s',
+        device.deviceName,
+        device.deviceType,
+        device.deviceId,
+      );
     }
   }
 
@@ -237,7 +262,11 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
     if (existingAccessory) {
       // the accessory already exists
       if (!this.config.options?.hide_device.includes(device.deviceId) && devices.statusCode === 100) {
-        this.log.info('Restoring existing accessory from cache:', existingAccessory.displayName, 'DeviceID:', device.deviceId);
+        this.log.info(
+          'Restoring existing accessory from cache: %s DeviceID: %s',
+          existingAccessory.displayName,
+          device.deviceId,
+        );
 
         // if you need to update the accessory.context then you should run `api.updatePlatformAccessories`. eg.:
         //existingAccessory.context.firmwareRevision = firmware;
@@ -253,8 +282,12 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       }
     } else if (!this.config.options?.hide_device.includes(device.deviceId)) {
       // the accessory does not yet exist, so we need to create it
-      this.log.info('Adding new accessory:', device.deviceName, device.deviceType, 'DeviceID:', device.deviceId);
-      this.log.debug('Registering new device:', device.deviceName, device.deviceType, '-', device.deviceId);
+      this.log.info(
+        'Adding new accessory: %s %s DeviceID: %s',
+        device.deviceName,
+        device.deviceType,
+        device.deviceId,
+      );
 
       // create a new accessory
       const accessory = new this.api.platformAccessory(`${device.deviceName} ${device.deviceType}`, uuid);
@@ -275,7 +308,79 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
       this.accessories.push(accessory);
     } else {
-      this.log.error(`Unable to Register new device: ${device.deviceName} ${device.deviceType} - ${device.deviceId}`);
+      this.log.error(
+        'Unable to Register new device: %s %s - %s',
+        device.deviceName,
+        device.deviceType,
+        device.deviceId,
+      );
+    }
+  }
+
+  private async createMeter(device: device, devices: deviceResponses) {
+    const uuid = this.api.hap.uuid.generate(
+      `${device.deviceName}-${device.deviceId}-${device.deviceType}-${device.hubDeviceId}`,
+    );
+
+    // see if an accessory with the same uuid has already been registered and restored from
+    // the cached devices we stored in the `configureAccessory` method above
+    const existingAccessory = this.accessories.find((accessory) => accessory.UUID === uuid);
+
+    if (existingAccessory) {
+      // the accessory already exists
+      if (!this.config.options?.hide_device.includes(device.deviceId) && devices.statusCode === 100) {
+        this.log.info(
+          'Restoring existing accessory from cache: %s DeviceID: %s',
+          existingAccessory.displayName,
+          device.deviceId,
+        );
+
+        // if you need to update the accessory.context then you should run `api.updatePlatformAccessories`. eg.:
+        //existingAccessory.context.firmwareRevision = firmware;
+        this.api.updatePlatformAccessories([existingAccessory]);
+        // create the accessory handler for the restored accessory
+        // this is imported from `platformAccessory.ts`
+        new Meter(this, existingAccessory, device);
+        this.log.debug(
+          `Meter UDID: ${device.deviceName}-${device.deviceId}-${device.deviceType}-${device.hubDeviceId}`,
+        );
+      } else {
+        this.unregisterPlatformAccessories(existingAccessory);
+      }
+    } else if (!this.config.options?.hide_device.includes(device.deviceId)) {
+      // the accessory does not yet exist, so we need to create it
+      this.log.info(
+        'Adding new accessory: %s %s DeviceID: %s',
+        device.deviceName,
+        device.deviceType,
+        device.deviceId,
+      );
+
+      // create a new accessory
+      const accessory = new this.api.platformAccessory(`${device.deviceName} ${device.deviceType}`, uuid);
+
+      // store a copy of the device object in the `accessory.context`
+      // the `context` property can be used to store any data about the accessory you may need
+      //accessory.context.firmwareRevision = firmware;
+      accessory.context.device = device;
+      // accessory.context.firmwareRevision = findaccessories.accessoryAttribute.softwareRevision;
+      // create the accessory handler for the newly create accessory
+      // this is imported from `platformAccessory.ts`
+      new Meter(this, accessory, device);
+      this.log.debug(
+        `Meter UDID: ${device.deviceName}-${device.deviceId}-${device.deviceType}-${device.hubDeviceId}`,
+      );
+
+      // link the accessory to your platform
+      this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
+      this.accessories.push(accessory);
+    } else {
+      this.log.error(
+        'Unable to Register new device: %s %s - %s',
+        device.deviceName,
+        device.deviceType,
+        device.deviceId,
+      );
     }
   }
 
@@ -291,7 +396,11 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
     if (existingAccessory) {
       // the accessory already exists
       if (!this.config.options?.hide_device.includes(device.deviceId) && devices.statusCode === 100) {
-        this.log.info('Restoring existing accessory from cache:', existingAccessory.displayName, 'DeviceID:', device.deviceId);
+        this.log.info(
+          'Restoring existing accessory from cache: %s DeviceID: %s',
+          existingAccessory.displayName,
+          device.deviceId,
+        );
 
         // if you need to update the accessory.context then you should run `api.updatePlatformAccessories`. eg.:
         //existingAccessory.context.firmwareRevision = firmware;
@@ -307,8 +416,12 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       }
     } else if (!this.config.options?.hide_device.includes(device.deviceId)) {
       // the accessory does not yet exist, so we need to create it
-      this.log.info('Adding new accessory:', device.deviceName, device.deviceType, 'DeviceID:', device.deviceId);
-      this.log.debug('Registering new device:', device.deviceName, device.deviceType, '-', device.deviceId);
+      this.log.info(
+        'Adding new accessory: %s %s DeviceID: %s',
+        device.deviceName,
+        device.deviceType,
+        device.deviceId,
+      );
 
       // create a new accessory
       const accessory = new this.api.platformAccessory(`${device.deviceName} ${device.deviceType}`, uuid);
@@ -329,7 +442,12 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
       this.accessories.push(accessory);
     } else {
-      this.log.error(`Unable to Register new device: ${device.deviceName} ${device.deviceType} - ${device.deviceId}`);
+      this.log.error(
+        'Unable to Register new device: %s %s - %s',
+        device.deviceName,
+        device.deviceType,
+        device.deviceId,
+      );
     }
   }