+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 All notable changes to this project will be documented in this file. This project uses [Semantic Versioning](https://semver.org/)
 
-## [Version 1.1.0](https://github.com/OpenWonderLabs/homebridge-switchbot-openapi/compare/v1.0.1...v1.1.0) (2021-01-XX)
+## [Version 1.1.0](https://github.com/OpenWonderLabs/homebridge-switchbot-openapi/compare/v1.0.1...v1.1.0) (2021-01-08)
 
 ### Changes
 
 - Allow for Hiding Devices based off of `DeviceID` instead of `DeviceType`.
-- Adds Support for [SwitchBot Curtain](https://www.switch-bot.com/products/switchbot-curtain).
+- Adds Support for [SwitchBot Meter](https://www.switch-bot.com/products/switchbot-meter).
+- Adds Beta Support for [SwitchBot Curtain](https://www.switch-bot.com/products/switchbot-curtain).
 
 ## [Version 1.0.1](https://github.com/OpenWonderLabs/homebridge-switchbot-openapi/compare/v1.0.0...v1.0.1) (2020-12-25)